Add files via upload
Retail Data Set
Context
The Challenge - One challenge of modelling retail data is the need to make decisions based on limited history. Holidays and select major events come once a year, and so does the chance to see how strategic decisions impacted the bottom line. In addition, markdowns are known to affect sales – the challenge is to predict which departments will be affected and to what extent. Content You are provided with historical sales data for 45 stores located in different regions - each store contains a number of departments. The company also runs several promotional markdown events throughout the year. These markdowns precede prominent holidays, the four largest of which are the Super Bowl, Labour Day, Thanksgiving, and Christmas. The weeks including these holidays are weighted five times higher in the evaluation than non-holiday weeks. Within the Excel Sheet, there are 3 Tabs – Stores, Features and Sales Stores Anonymized information about the 45 stores, indicating the type and size of store

Features Contains additional data related to the store, department, and regional activity for the given dates. • Store - the store number • Date - the week • Temperature - average temperature in the region • Fuel Price - cost of fuel in the region • MarkDown1-5 - anonymized data related to promotional markdowns. Markdown data is only available after Nov 2011, and is not available for all stores all the time. Any missing value is marked with an NA • CPI - the consumer price index • Unemployment - the unemployment rate • Is Holiday - whether the week is a special holiday week Sales Historical sales data, which covers to 2010-02-05 to 2012-11-01. Within this tab you will find the following fields: • Store - the store number • Dept - the department number • Date - the week • Weekly Sales- sales for the given department in the given store • Is Holiday - whether the week is a special holiday week

The Task

Predict the department-wide sales for each store for the following year
Model the effects of markdowns on holiday weeks
Provide recommended actions based on the insights drawn, with prioritization placed on largest business impact